(PC-17103) auth: move useLogoutRoutine to a separate file
