+++ src/features/auth/__mocks__/AuthContext.tsx
@@ -6,7 +6,3 @@ export const useAuthContext = jest.fn().mockReturnValue({
   setIsLoggedIn,
   isLoggedIn: false,
 }) as jest.MockedFunction<typeof actualUseAuthContext>
-
-export const signOut = jest.fn()
-
-export const useLogoutRoutine = jest.fn(() => signOut)
