+++ src/features/auth/logout/__mocks__/useLogoutRoutine.ts
@@ -1,0 +1,1 @@
+export const useLogoutRoutine = jest.fn(() => jest.fn())
