+++ src/features/auth/suspendedAccount/SuspendedAccount/tests/SuspendedAccount.native.test.tsx
@@ -19,7 +19,7 @@ jest.mock('features/auth/suspendedAccount/SuspendedAccount/useAccountSuspensionD
 jest.mock('features/navigation/helpers')
 
 const mockSignOut = jest.fn()
-jest.mock('features/auth/AuthContext', () => ({
+jest.mock('features/auth/logout/useLogoutRoutine', () => ({
   useLogoutRoutine: jest.fn(() => mockSignOut.mockResolvedValueOnce(jest.fn())),
 }))
