+++ src/features/profile/pages/AfterChangeEmailValidationBuffer/AfterChangeEmailValidationBuffer.tsx
@@ -4,7 +4,8 @@ import { useMutation } from 'react-query'
 
 import { api } from 'api/api'
 import { ChangeBeneficiaryEmailBody } from 'api/gen'
-import { useAuthContext, useLogoutRoutine } from 'features/auth/AuthContext'
+import { useAuthContext } from 'features/auth/AuthContext'
+import { useLogoutRoutine } from 'features/auth/logout/useLogoutRoutine'
 import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator/types'
 import { isTimestampExpired } from 'libs/dates'
 import { LoadingPage } from 'ui/components/LoadingPage'
