+++ src/features/profile/pages/AfterChangeEmailValidationBuffer/__tests__/AfterChangeEmailValidationBuffer.native.test.tsx
@@ -42,8 +42,11 @@ jest.mock('ui/components/snackBar/SnackBarContext', () => ({
 }))
 
 const mockSignOut = jest.fn()
-jest.mock('features/auth/AuthContext', () => ({
+jest.mock('features/auth/logout/useLogoutRoutine', () => ({
   useLogoutRoutine: () => mockSignOut,
+}))
+
+jest.mock('features/auth/AuthContext', () => ({
   useAuthContext: jest.fn().mockReturnValue({
     isLoggedIn: true,
   }),
